AuditLog: timeSpent updated from Long to Integer

+++ webservice-global/src/main/java/com/sos/joc/classes/audit/IAuditLog.java
@@ -14,7 +14,7 @@ public interface IAuditLog {
     
     public String getJobschedulerId();
     
-    public Long getTimeSpent();
+    public Integer getTimeSpent();
     
     public String getTicketLink();
 }
