+++ webservice-global/src/main/java/com/sos/joc/classes/audit/JobSchedulerCommandAudit.java
@@ -66,7 +66,7 @@ public class JobSchedulerCommandAudit extends JobschedulerCommands implements IA
 
     @Override
     @JsonIgnore
-    public Long getTimeSpent() {
+    public Integer getTimeSpent() {
         return null;
     }
